% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_generate_lvm.r
\name{generateLV.irt}
\alias{generateLV.irt}
\title{method for all IRT model}
\usage{
\method{generateLV}{irt}(info)
}
\description{
method for all IRT model
}
\examples{

lvmodel <- "gpcm"
ipar <- genIRTpar(20, ncat = 3, 2, lvmodel)
eta <- MASS::mvrnorm(100, rep(0, 2), matrix(c(1,0,0,1),ncol=2))
generateLV.irt(lvmodel, eta, ipar)

}
