% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_generate_true_eta.r
\name{genTrueEta}
\alias{genTrueEta}
\title{S3 generic for individual latent scores data generation}
\usage{
genTrueEta(Data, ...)
}
\description{
S3 generic for individual latent scores data generation
}
