% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPS2022-package.r
\docType{package}
\name{IMPS2022-package}
\alias{IMPS2022-package}
\alias{IMPS2022}
\title{IMPS2022 proceedings}
\description{
The IMPS2022 package contains the data generation codes for
 IMPS2022 proceedings.
}
\details{
The 'IMPS2022' package.
}
\references{
Sales, A. C., & Pane, J. F. (2019). The role of mastery learning in an intelligent tutoring system: Principal stratification on a latent variable. The Annals of Applied Statistics, 13(1), 420-443.
}
\author{
Sooyong Lee
}
\keyword{package}
