% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_make_data.r
\name{makeDat}
\alias{makeDat}
\title{Generate Fully Latent Principal Stratification data}
\usage{
makeDat(N, R2Y, R2eta, omega, tau0, tau1, lambda, nsec, lvmodel)
}
\arguments{
\item{N}{a numeric indicating sample size.}

\item{R2Y}{a numeric indicating predictive power of covariates.}

\item{R2eta}{a numeric indicating Predictive power of latent variable}

\item{lambda}{a numeric indicating the mean of Worked problems/person.
(extent to which covariates predict eta).}

\item{nsec}{a numeric indicating the number of maximum sections given to
students.}

\item{lvmodel}{a character specifying a type of latent variable model.}

\item{linear}{a logical}

\item{ydist}{a character}

\item{nfac}{a numeric indicating the number of latent factors}
}
\value{
a list containing all the data related to population values and running FLPS.
}
\description{
\code{\link{makeDat}} is a function for generating a data based on the given
information.
}
\examples{
sdat <- makeDat(
  N = 100,
  R2Y = 0.2,
  R2eta = 0.5,
  linear = T,
  ydist = "n",
  lambda = .6,
  nsec = 10,
  nfac = 1,
  lvmodel = "2PL"
)

}
